+++ flag/service/installation/installation.go
@@ -1,10 +1,9 @@
 package installation
 
 type Installation struct {
-	Customer          string
-	Name              string
-	Pipeline          string
-	Region            string
-	Registry          string
-	IngressAPIVersion string
+	Customer string
+	Name     string
+	Pipeline string
+	Region   string
+	Registry string
 }