+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improve HasPrometheusAgent function to ignore prometheus-agent scraping targets.
+- Do not reconcile service monitors in CAPO MCs.
 
 ## [4.12.0] - 2022-11-25