+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bump alpine from 3.16.3 to 3.17.0
+- Do not reconcile service monitors in CAPO MCs.
+
+### Removed
+
+- Remove deprecated ingress v1beta1 (only used on kvm).
 
 ## [4.12.0] - 2022-11-25